dereference:true 처리
dereference  : By default, the references to the values (current, init, ...) of the property will be stored as references on the object. When disposing this object, the references will not be deleted. Setting the dereference key to true tells the property system to delete all connections made by this property on dispose. This can be necessary for disconnecting DOM objects to allow the garbage collector to work properly.